refactor: `MapStringConverter` 패키지 이동
