+++ backend/src/main/java/harustudy/backend/common/MapStringConverter.java
@@ -1,4 +1,4 @@
-package harustudy.backend.content.domain;
+package harustudy.backend.common;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
